refactor(engine): move StreamProcessorListener to streamprocessor package
