+++ engine/src/main/java/io/camunda/zeebe/streamprocessor/StreamProcessorListener.java
@@ -5,11 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.processing.streamprocessor;
+package io.camunda.zeebe.streamprocessor;
 
 import io.camunda.zeebe.engine.api.TypedRecord;
 import io.camunda.zeebe.logstreams.log.LoggedEvent;
-import io.camunda.zeebe.streamprocessor.StreamProcessor;
 
 /**
  * A listener for the {@link StreamProcessor}. Allows retrieving insides of the processing and
